+++ examples/react-router-v7/app/entry.server.tsx
@@ -1,16 +1,16 @@
-import { PassThrough } from "node:stream";
-
-import type { AppLoadContext, EntryContext } from "@remix-run/node";
-import { createReadableStreamFromReadable } from "@remix-run/node";
-import { RemixServer } from "@remix-run/react";
-import { isbot } from "isbot";
-import { renderToPipeableStream } from "react-dom/server";
 import {
   createNonce,
   createSecureHeaders,
   mergeHeaders,
 } from "@mcansh/http-helmet";
 import { NonceProvider } from "@mcansh/http-helmet/react";
+import { createReadableStreamFromReadable } from "@react-router/node";
+import { isbot } from "isbot";
+import { PassThrough } from "node:stream";
+import type { RenderToPipeableStreamOptions } from "react-dom/server";
+import { renderToPipeableStream } from "react-dom/server";
+import type { AppLoadContext, EntryContext } from "react-router";
+import { ServerRouter } from "react-router";
 
 const ABORT_DELAY = 5_000;
 
@@ -18,43 +18,39 @@ export default function handleRequest(
   request: Request,
   responseStatusCode: number,
   responseHeaders: Headers,
-  remixContext: EntryContext,
-  loadContext: AppLoadContext
+  routerContext: EntryContext,
+  _loadContext: AppLoadContext
 ) {
-  let callback = isbot(request.headers.get("user-agent"))
-    ? "onAllReady"
-    : "onShellReady";
-
-  let nonce = createNonce();
-  let secureHeaders = createSecureHeaders({
+  const nonce = createNonce();
+  const secureHeaders = createSecureHeaders({
     "Content-Security-Policy": {
-      defaultSrc: ["'self'"],
-      scriptSrc: ["'self'", `'nonce-${nonce}'`],
-      connectSrc: [
-        "'self'",
-        ...(process.env.NODE_ENV === "development" ? ["ws:", ""] : []),
-      ],
-    },
-    "Strict-Transport-Security": {
-      maxAge: 31536000,
-      includeSubDomains: true,
-      preload: true,
+      "script-src": ["'self'", `'nonce-${nonce}'`],
     },
   });
 
   return new Promise((resolve, reject) => {
     let shellRendered = false;
+    let userAgent = request.headers.get("user-agent");
+
+    // Ensure requests from bots and SPA Mode renders wait for all content to load before responding
+    // https://react.dev/reference/react-dom/server/renderToPipeableStream#waiting-for-all-content-to-load-for-crawlers-and-static-generation
+    let readyOption: keyof RenderToPipeableStreamOptions =
+      (userAgent && isbot(userAgent)) || routerContext.isSpaMode
+        ? "onAllReady"
+        : "onShellReady";
+
     const { pipe, abort } = renderToPipeableStream(
       <NonceProvider nonce={nonce}>
-        <RemixServer
-          context={remixContext}
+        <ServerRouter
+          context={routerContext}
           url={request.url}
           abortDelay={ABORT_DELAY}
+          nonce={nonce}
         />
       </NonceProvider>,
       {
         nonce,
-        [callback]() {
+        [readyOption]() {
           shellRendered = true;
           const body = new PassThrough();
           const stream = createReadableStreamFromReadable(body);