+++ examples/vite/vite.config.ts
@@ -9,7 +9,6 @@ export default defineConfig({
   plugins: [
     remix({
       future: {
-        unstable_singleFetch: true,
         v3_fetcherPersist: true,
         v3_relativeSplatPath: true,
         v3_throwAbortReason: true,