+++ packages/http-helmet/src/helmet.ts
@@ -1,8 +1,6 @@
-import { RequireExactlyOne } from "type-fest";
-import {
-  ContentSecurityPolicy,
-  createContentSecurityPolicy,
-} from "./rules/content-security-policy.js";
+import type { RequireOneOrNone } from "type-fest";
+import { createContentSecurityPolicy } from "./rules/content-security-policy.js";
+import type { PublicContentSecurityPolicy } from "./rules/content-security-policy.js";
 import {
   createPermissionsPolicy,
   PermissionsPolicy,
@@ -12,8 +10,8 @@ import {
   StrictTransportSecurity,
 } from "./rules/strict-transport-security.js";
 
+export type { PublicContentSecurityPolicy as ContentSecurityPolicy };
 export { createContentSecurityPolicy } from "./rules/content-security-policy.js";
-export type { ContentSecurityPolicy } from "./rules/content-security-policy.js";
 export { createPermissionsPolicy } from "./rules/permissions.js";
 export type { PermissionsPolicy } from "./rules/permissions.js";
 export { createStrictTransportSecurity } from "./rules/strict-transport-security.js";
@@ -102,20 +100,20 @@ type BaseSecureHeaders = {
   "Cross-Origin-Resource-Policy"?: "same-site" | "same-origin" | "cross-origin";
 };
 
-export type CreateSecureHeaders = RequireExactlyOne<
+export type CreateSecureHeaders = RequireOneOrNone<
   {
     /**
      * @description Content Security Policy (CSP) is an added layer of security that helps to detect and mitigate certain types of attacks, including Cross-Site Scripting (XSS) and data injection attacks. These attacks are used for everything from data theft, to site defacement, to malware distribution.
      * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP
      */
 
-    "Content-Security-Policy"?: ContentSecurityPolicy;
+    "Content-Security-Policy"?: PublicContentSecurityPolicy;
 
     /**
      * @description The HTTP Content-Security-Policy-Report-Only response header allows web developers to experiment with policies by monitoring (but not enforcing) their effects. These violation reports consist of JSON documents sent via an HTTP POST request to the specified URI defined in a Reporting-Endpoints HTTP response header.
      * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy-Report-Only
      */
-    "Content-Security-Policy-Report-Only"?: ContentSecurityPolicy;
+    "Content-Security-Policy-Report-Only"?: PublicContentSecurityPolicy;
   },
   "Content-Security-Policy" | "Content-Security-Policy-Report-Only"
 > &
@@ -129,14 +127,14 @@ export function createSecureHeaders(options: CreateSecureHeaders) {
     options["Content-Security-Policy-Report-Only"]
   ) {
     throw new Error(
-      "createSecureHeaders: Content-Security-Policy and Content-Security-Policy-Report-Only cannot be set at the same time",
+      "createSecureHeaders: Content-Security-Policy and Content-Security-Policy-Report-Only cannot be set at the same time"
     );
   }
 
   if (options["Content-Security-Policy"]) {
     headers.set(
       "Content-Security-Policy",
-      createContentSecurityPolicy(options["Content-Security-Policy"]),
+      createContentSecurityPolicy(options["Content-Security-Policy"])
     );
   }
 
@@ -144,8 +142,8 @@ export function createSecureHeaders(options: CreateSecureHeaders) {
     headers.set(
       "Content-Security-Policy-Report-Only",
       createContentSecurityPolicy(
-        options["Content-Security-Policy-Report-Only"],
-      ),
+        options["Content-Security-Policy-Report-Only"]
+      )
     );
   }
 
@@ -156,14 +154,14 @@ export function createSecureHeaders(options: CreateSecureHeaders) {
   if (options["Permissions-Policy"]) {
     headers.set(
       "Permissions-Policy",
-      createPermissionsPolicy(options["Permissions-Policy"]),
+      createPermissionsPolicy(options["Permissions-Policy"])
     );
   }
 
   if (options["Strict-Transport-Security"]) {
     headers.set(
       "Strict-Transport-Security",
-      createStrictTransportSecurity(options["Strict-Transport-Security"]),
+      createStrictTransportSecurity(options["Strict-Transport-Security"])
     );
   }
 
@@ -178,21 +176,21 @@ export function createSecureHeaders(options: CreateSecureHeaders) {
   if (options["Cross-Origin-Embedder-Policy"]) {
     headers.set(
       "Cross-Origin-Embedder-Policy",
-      options["Cross-Origin-Embedder-Policy"],
+      options["Cross-Origin-Embedder-Policy"]
     );
   }
 
   if (options["Cross-Origin-Opener-Policy"]) {
     headers.set(
       "Cross-Origin-Opener-Policy",
-      options["Cross-Origin-Opener-Policy"],
+      options["Cross-Origin-Opener-Policy"]
     );
   }
 
   if (options["Cross-Origin-Resource-Policy"]) {
     headers.set(
       "Cross-Origin-Resource-Policy",
-      options["Cross-Origin-Resource-Policy"],
+      options["Cross-Origin-Resource-Policy"]
     );
   }