+++ packages/http-helmet/src/rules/content-security-policy.ts
@@ -1,43 +1,50 @@
 import { LiteralUnion, KebabCasedProperties } from "type-fest";
-import { QuotedSource, dashify, isQuoted } from "../utils.js";
+import { QuotedSource, isQuoted } from "../utils.js";
+import { kebabCase } from "change-case";
+
+type CspSetting = Array<LiteralUnion<QuotedSource, string> | undefined>;
 
 type ContentSecurityPolicyCamel = {
-  childSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  connectSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  defaultSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  fontSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  frameSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  imgSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  manifestSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  mediaSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  objectSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  prefetchSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  scriptSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  scriptSrcElem?: Array<LiteralUnion<QuotedSource, string>>;
-  scriptSrcAttr?: Array<LiteralUnion<QuotedSource, string>>;
-  styleSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  styleSrcElem?: Array<LiteralUnion<QuotedSource, string>>;
-  styleSrcAttr?: Array<LiteralUnion<QuotedSource, string>>;
-  workerSrc?: Array<LiteralUnion<QuotedSource, string>>;
-  baseUri?: Array<LiteralUnion<QuotedSource, string>>;
-  sandbox?: Array<LiteralUnion<QuotedSource, string>>;
-  formAction?: Array<LiteralUnion<QuotedSource, string>>;
-  frameAncestors?: Array<LiteralUnion<QuotedSource, string>>;
-  navigateTo?: Array<LiteralUnion<QuotedSource, string>>;
-  reportUri?: Array<LiteralUnion<QuotedSource, string>>;
-  reportTo?: Array<LiteralUnion<QuotedSource, string>>;
-  requireSriFor?: Array<LiteralUnion<QuotedSource, string>>;
-  requireTrustedTypesFor?: Array<LiteralUnion<QuotedSource, string>>;
-  trustedTypes?: Array<LiteralUnion<QuotedSource, string>>;
+  childSrc?: CspSetting;
+  connectSrc?: CspSetting;
+  defaultSrc?: CspSetting;
+  fontSrc?: CspSetting;
+  frameSrc?: CspSetting;
+  imgSrc?: CspSetting;
+  manifestSrc?: CspSetting;
+  mediaSrc?: CspSetting;
+  objectSrc?: CspSetting;
+  prefetchSrc?: CspSetting;
+  scriptSrc?: CspSetting;
+  scriptSrcElem?: CspSetting;
+  scriptSrcAttr?: CspSetting;
+  styleSrc?: CspSetting;
+  styleSrcElem?: CspSetting;
+  styleSrcAttr?: CspSetting;
+  workerSrc?: CspSetting;
+  baseUri?: CspSetting;
+  sandbox?: CspSetting;
+  formAction?: CspSetting;
+  frameAncestors?: CspSetting;
+  navigateTo?: CspSetting;
+  reportUri?: CspSetting;
+  reportTo?: CspSetting;
+  requireSriFor?: CspSetting;
+  requireTrustedTypesFor?: CspSetting;
+  trustedTypes?: CspSetting;
   upgradeInsecureRequests?: boolean;
 };
 
 type ContentSecurityPolicyKebab =
   KebabCasedProperties<ContentSecurityPolicyCamel>;
 
-export interface ContentSecurityPolicy
-  extends ContentSecurityPolicyCamel,
-    ContentSecurityPolicyKebab {}
+type ContentSecurityPolicy =
+  | ContentSecurityPolicyCamel
+  | ContentSecurityPolicyKebab;
+
+export type PublicContentSecurityPolicy = Parameters<
+  typeof createContentSecurityPolicy
+>[0];
 
 let reservedCSPKeywords = new Set([
   "self",
@@ -47,42 +54,59 @@ let reservedCSPKeywords = new Set([
 ]);
 
 export function createContentSecurityPolicy(
+  settings: ContentSecurityPolicyCamel,
+): string;
+export function createContentSecurityPolicy(
+  settings: ContentSecurityPolicyKebab,
+): string;
+export function createContentSecurityPolicy(
   settings: ContentSecurityPolicy,
 ): string {
-  let { upgradeInsecureRequests, ...rest } = settings;
-  let policy: Array<string> = [];
-  let seenKeys: Set<string> = new Set();
+  let { "upgrade-insecure-requests": upgradeInsecureRequests, ...rest } =
+    Object.entries(settings).reduce<ContentSecurityPolicyKebab>(
+      (acc, [key, value]) => {
+        let kebab = kebabCase(key) as keyof ContentSecurityPolicyKebab;
+        if (acc[kebab]) {
+          throw new Error(
+            `[createContentSecurityPolicy]: The key "${key}" was specified in camelCase and kebab-case.`,
+          );
+        }
+        // @ts-expect-error - hush
+        acc[kebab] = value;
+        return acc;
+      },
+      {},
+    );
 
-  if (settings.upgradeInsecureRequests) {
+  let policy: Array<string> = [];
+
+  if (upgradeInsecureRequests) {
     policy.push("upgrade-insecure-requests");
   }
 
-  for (let [originalKey, values] of Object.entries(rest)) {
-    let key = dashify(originalKey);
-    if (seenKeys.has(key)) {
-      throw new Error(
-        `[createContentSecurityPolicy]: The key "${originalKey}" was specified more than once.`,
-      );
-    }
-
+  for (let [key, values] of Object.entries(rest)) {
     let allowedValuesSeen: Set<string> = new Set();
 
     if (!Array.isArray(values)) {
       throw new Error(
-        `[createContentSecurityPolicy]: The value of the "${originalKey}" must be array of strings.`,
+        `[createContentSecurityPolicy]: The value of the "${key}" must be array of strings.`,
       );
     }
 
-    values.forEach((allowedValue) => {
+    let definedValues = values.filter(
+      (v): v is string => typeof v !== "undefined",
+    );
+
+    definedValues.forEach((allowedValue) => {
       if (typeof allowedValue !== "string") {
         throw new Error(
-          `[createContentSecurityPolicy]: The value of the "${originalKey}" contains a non-string, which is not supported.`,
+          `[createContentSecurityPolicy]: The value of the "${key}" contains a non-string, which is not supported.`,
         );
       }
 
       if (allowedValuesSeen.has(allowedValue)) {
         throw new Error(
-          `[createContentSecurityPolicy]: The value of the "${originalKey}" contains duplicates, which it shouldn't.`,
+          `[createContentSecurityPolicy]: The value of the "${key}" contains duplicates, which it shouldn't.`,
         );
       }
 
@@ -95,8 +119,13 @@ export function createContentSecurityPolicy(
       allowedValuesSeen.add(allowedValue);
     });
 
-    policy.push(`${key} ${values.filter(Boolean).join(" ")}`);
-    seenKeys.add(key);
+    if (definedValues.length === 0) {
+      throw new Error(
+        `[createContentSecurityPolicy]: key "${key}" has no defined options`,
+      );
+    }
+
+    policy.push(`${key} ${definedValues.join(" ")}`);
   }
 
   return policy.join("; ");